updated new files in this peoject, added css to this project, updated user table in database

+++ delete_task.php
@@ -2,18 +2,20 @@
 require('authenticate.php');
 require('connect.php');
 
+if ($_SESSION['user_type'] !== 'admin') {
+    echo "You are not authorized to access this page.";
+    exit;
+}
+
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_POST['task_id'])) {
     try {
-        // Delete task
         $task_id = filter_input(INPUT_POST, 'task_id', FILTER_SANITIZE_NUMBER_INT);
         $query = "DELETE FROM task WHERE task_id = :task_id";
         $statement = $db->prepare($query);
         $statement->bindValue(':task_id', $task_id, PDO::PARAM_INT);
 
-        // Execute the deletion query
         $statement->execute();
 
-        // Check if any rows were affected
         if ($statement->rowCount() > 0) {
             header("Location: index.php");
             exit;
